Move calculation of new images to a custom hook
